+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [0.5.4] - 2019-09-05
+### Fixed
+- The `maximum content length` was being superseded by a NGINX content
+  length limiter.
+- Word add-in bug.
+
 ## [0.5.3] - 2019-09-04
 ### Added
 - The `exit url` specifier under `metadata`.